Remove extra if and add flag for enabling project setup in setup.sh

+++ setup.sh
@@ -8,6 +8,7 @@ CUDA_HOME=/usr/local/cuda-10.2      # Must be >=Cuda-10.1 <Cuda-11.0
 BUILD_FRESH=true     # Deletes the existing repo, clones a fresh repo and builds from that
 BUILD_KALDI=true
 BUILD_ESPNET=true
+SETUP_PROJECT=true
 
 # Install Kaldi
 KALDI_ROOT=$(pwd)/kaldi
@@ -33,7 +34,6 @@ then
     popd
     pushd kaldi/src
     ./configure --use-cuda=no    # ESPNet only uses Kaldi feature extraction which is CPU bound
-    if 
     # $ ./configure --openblas-root=../tools/OpenBLAS/install --cudatk-dir=$CUDA_HOME    # If you want to build Kaldai with CUDA and BLAS
     make -j$(nproc) clean depend
     make -j$(nproc)
@@ -63,9 +63,14 @@ then
         ./venv/bin/pip install ipykernel
     elif [ "$WHICH_PYTHON_ENV" == "anaconda" ]
     then
+        CONDA_ENV_NAME="espnet"
+        if $BUILD_FRESH
+        then
+            conda env remove --name=$CONDA_ENV_NAME
+        fi
         CONDA_TOOLS_DIR=$(dirname ${CONDA_EXE})/..
-        ./setup_anaconda.sh ${CONDA_TOOLS_DIR} espnet 3
-        conda install ipykernel --name=espnet-conda-venv
+        ./setup_anaconda.sh ${CONDA_TOOLS_DIR} $CONDA_ENV_NAME 3
+        conda install -y ipykernel --name=$CONDA_ENV_NAME
     elif [ "$WHICH_PYTHON_ENV" == "sys" ]
     then
         ./setup_python.sh $(command -v python3)
@@ -85,28 +90,38 @@ then
 fi
 
 ESPNET_WSJ=$ESPNET_ROOT/egs/wsj/asr1
-
-# Project Specific Setup
-rm -rf $ESPNET_WSJ/conf $ESPNET_WSJ/run.sh
-ln -s $(pwd)/wsj_asr1/conf $ESPNET_WSJ
-ln -s $(pwd)/wsj_asr1/run.sh $ESPNET_WSJ
-
-# Pull Model
 MODEL="pretrained-transformer-model"
 MODEL_PATH=$(pwd)/$MODEL
 
-# Pretrained Models
-if [ $MODEL == "pretrained-transformer-model" ]
+if $SETUP_PROJECT
 then
-    $ESPNET_ROOT/utils/download_from_google_drive.sh "https://drive.google.com/open?id=1Az-4H25uwnEFa4lENc-EKiPaWXaijcJp" $MODEL_PATH tar.gz
-    mkdir -p $ESPNET_WSJ/data/train_si284
-    ln -s $MODEL_PATH/data/lang_1char $ESPNET_WSJ/data
-    ln -s $MODEL_PATH/data/train_si284/cmvn.ark $ESPNET_WSJ/data/train_si284
-    mkdir -p $ESPNET_WSJ/exp
-    ln -s $MODEL_PATH/exp/train_rnnlm_pytorch_lm_word65000 $ESPNET_WSJ/exp
-    ln -s $MODEL_PATH/exp/train_si284_pytorch_train_no_preprocess $ESPNET_WSJ/exp
-elif [ $MODEL == "pretrained-rnn-model" ]
-then
-    $ESPNET_ROOT/utils/download_from_google_drive.sh "https://drive.google.com/u/0/uc?id=1Az-4H25uwnEFa4lENc-EKiPaWXaijcJp&export=download" $MODEL_PATH tar.gz
-    # Haven't pulled and tested this yet
+    # Project Specific Setup
+    rm -rf $ESPNET_WSJ/conf $ESPNET_WSJ/run.sh
+    ln -s $(pwd)/wsj_asr1/conf $ESPNET_WSJ
+    ln -s $(pwd)/wsj_asr1/run.sh $ESPNET_WSJ
+
+    # Pretrained Models
+    if [ $MODEL == "pretrained-transformer-model" ]
+    then
+        if [ ! -d $MODEL_PATH ]
+        then
+            $ESPNET_ROOT/utils/download_from_google_drive.sh "https://drive.google.com/open?id=1Az-4H25uwnEFa4lENc-EKiPaWXaijcJp" $MODEL_PATH tar.gz
+        fi
+        mkdir -p $ESPNET_WSJ/data/train_si284
+        ln -s $MODEL_PATH/data/lang_1char $ESPNET_WSJ/data
+        ln -s $MODEL_PATH/data/train_si284/cmvn.ark $ESPNET_WSJ/data/train_si284
+        mkdir -p $ESPNET_WSJ/exp
+        ln -s $MODEL_PATH/exp/train_rnnlm_pytorch_lm_word65000 $ESPNET_WSJ/exp
+        ln -s $MODEL_PATH/exp/train_si284_pytorch_train_no_preprocess $ESPNET_WSJ/exp
+    elif [ $MODEL == "pretrained-rnn-model" ]
+    then
+        if [ ! -d $MODEL_PATH ]
+        then
+            $ESPNET_ROOT/utils/download_from_google_drive.sh "https://drive.google.com/u/0/uc?id=1Az-4H25uwnEFa4lENc-EKiPaWXaijcJp&export=download" $MODEL_PATH tar.gz
+        fi
+        echo "Have not previously pulled or tested this yet. Setup is still required"
+    elif [ $MODEL == "wsj_asr1" ]
+    then
+        echo "We have not trained this model to completion yet"
+    fi
 fi
